Allow for extensions on a per filesystem view basis [ALREADY PR]
Resolves #1789

In our scenario we have a a globally installed (i.e., read-only) set of
packages from which each user should then be able to add a subset to
local filesystem views. However, the current extension mechanism only
allows for one set of extensions to be activated for any given extendee.

One way to circumvent this issue is to use a module system, but this way
environment variables tend to get very long after a short while.
Furthermore, since an error is raised when two files have the same
target in the view, the user can immediately spot when multiple specs
from one package are loaded (an error that sill creeps up from time to
time).

In order to support multiple (and maybe mutually exclusive)
configurations of extensions for a given extendee, the following changes
were made. The aim is to silently activate any specified extensions when
adding them to the filesystem view; no actions from the user required.

If an extension is globally activated, it will not be linked again by the
view. Note that the view might break, however, if the extension gets
(globally) deactivated later on.

TL;DR: Adding python packages built with py-setuptools to fileystem
views without globally activating them should "just work".

* `DirectoryLayout`/`Package`: All extension-related command accept a
`path_view` argument restricting extension operations to the given
filesystem view.

* `Package`: `{do_,}{de,}activate`-functions also accept a `path_view`
argument. If `path_view` is `not None`, the necessary extension files
are linked into the view instead of the extendee's install directory.

* `LinkTree`: `merge`-method now accepts a link argument to allow for
both symlinks and hardlinks.

* `view`: All linking operations have been refactored to use `LinkTree`s
instead of custom code.

* `view`: Check extensions prior to linking packages, activate
extensions in filesystem view afterwards.

* `view statlink`: The `--verbose` option now prints the `short_spec` of
packages linked.

* `view statlink`: `status` and `remove` now accept `--all` argument to
display the status of/remove all packages in the filesystem view.

* `view statlink`: Now uses `display_specs`-function for prettier
output.

* `view remove`: Only removes packages specified by user, packages
depending on those packages and packages that only user-specified
packages depend on.