Improve 'buildcache install/create'
* Add -w/--without-dependencies to suppress dependency resolution.

* Add -j/--jobs to install several files from buildcache in parallel
  (defaults to 1).

The current buildcache implementation lacks the ability to install
packages/create tarballs en bulk:

* Dependencies are ALWAYS checked for and included
* the package index is updated after EVERY installed package instead of
  once after everything is installed

In order to reduce build time of the container we want to pre-compute
all hashes to be installed from buildcache in advance (i.e. we do not
want spack to perform additional checks) and we want this installation
to be possible with more than one core (decompressing is cpu-intensive
and therefore not completely IO-bound).

The same goes for creating tarballs that are put into the buildcache.

Hence this ChangeSet introduces these two possibilities.

By default the behavior remains the same.

In theory parallelization could also be achived by using tools such as
xargs and spawning multiple python processes, but I think it would be
nicer to have this possiblity out of the box in spack (in the same way
as you can install packages with multiple cores).

Change-Id: I97b1c14bf7563ac9d68b904b0c38eeb08b8d210e