Add multiprocessing to buildcache command
* Add `--only {package,dependencies}` to control whether to
  create/install the specified packages/dependencies or both (the
  default).

* Add -j/--jobs to install/create several package from/to buildcache in
  parallel (defaults to 1).

* Applied some flake8-fixes

Reasoning:

The current buildcache implementation lacks the ability to install
packages/create tarballs en bulk:

* Dependencies are *always* checked for and included
* the package index is updated after EVERY installed package instead of
  once after everything is installed

In order to reduce build time of our spack-based singularity container
we want to pre-compute all hashes to be installed from buildcache in
advance (i.e. we do not want spack to perform additional checks) and we
want this installation to be possible with more than one core
(decompressing is cpu-intensive and therefore not completely IO-bound).

The same goes for creating tarballs that are put into the buildcache
after container creation.

Hence this commit introduces these two possibilities.

By default the behavior remains the same, i.e. spack fails after
encountering its first error. When using several processes we gather all
occurred errors and print them afterwards.

The only (minor) difference is that since it is not possible to raise
exceptions in multiprocessing.Pool-workers we have to convert all
encountered SpackErrors to strings and print them manually. In the
single-threaded case we re-raise the exception.

In theory parallelization could also be achieved by using tools such as
xargs and spawning multiple python processes, but I think it would be
nicer to have this possibility out of the box in spack (in the same way
as you can install packages with multiple cores).

Change-Id: I7b52730d37bd68692c8f7bf03f87d84a58272f49