Fix parsing of key=value pairs
Fix parsing key-value pairs being too greedy and only working if the
key=value pair is the last entry in the spec.

Example:
```
$ spack spec "py-matplotlib backend=agg target=x86_64 +image ^python@3.7.5"
==> Error: invalid values for variant "backend" in package "py-matplotlib": ['agg target=x86_64 +image ^python@3.7.5']
```

The error is that the parsing is actually happening in two stages:
* First, `shlex.split` is used to tokenize the input string.
* Second, each token is parsed individually. A `key=value`-pair is
  checked to contain an `=`. Everything leading up to it is the key, the
  remainder is the value. The value is ensured to start with a
  non-whitespace character.

Parsing has to be done this way because the current lexer implementation
expects input to only contain single tokens and is not able to split an
input stream into several tokens. While it has basic capabilities to
switch parsing modes,it is currently only able to do so once once (see
`lib/spack/spack/parse.py:lex_word`).

In order to accommodate for this, command line arguments are quoted and
joined to form a single string (in
`lib/spack/spack/cmd/__init__.py:parse_specs`). If the whole spec is
given as a single argument, it gets double-escaped and the lexer is
unable to handle more than one token occurring in a single input word.

Fix (which is more of a workaround): Check if the list specs to parse
only contains one item and avoid quoting it.

A more long term solution would be a more robust parser implementation
that does not rely on input being tokenized via `shlex.split`
beforehand.

Fixes: #13752

Change-Id: I2314bbddff385fe5ccf6a39260332ec842d1b0cf